DLLMySQL: Added Login & Withdraw functions
Added Login function which checks for corresponding data from the database and Withdraw that first checks the corresponding ID's balance and then checks if the amount to be withdrawn is equal or less than the total balance. If true, does the math and updates the total balance to its new value.